# openapi.yaml
openapi: 3.0.1
info:
    title: Pratica 4 API
    description: Pratica 4 API documentation
    termsOfService: ''
    contact:
        name: ''
        url: ''
        email: 'pedpao@pedpao.com'
    license:
        name: unlicensed
        url: ''
    version: 0.0.1
servers:
    - url: http://localhost:8080
      description: Generated server url
paths:
...
    /api/items-assinaturas/{id}:
        get:
            tags:
                - items-assinatura-resource
            operationId: getItemsAssinatura
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                '200':
                    description: OK
                    content:
                        '*/*':
                            schema:
                                $ref: '#/components/schemas/ItemsAssinaturaDTO'
        put:
            tags:
                - items-assinatura-resource
            operationId: updateItemsAssinatura
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ItemsAssinaturaDTO'
                required: true
            responses:
                '200':
                    description: OK
                    content:
                        '*/*':
                            schema:
                                $ref: '#/components/schemas/ItemsAssinaturaDTO'
        delete:
            tags:
                - items-assinatura-resource
            operationId: deleteItemsAssinatura
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                '200':
                    description: OK
        patch:
            tags:
                - items-assinatura-resource
            operationId: partialUpdateItemsAssinatura
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ItemsAssinaturaDTO'
                    application/merge-patch+json:
                        schema:
                            $ref: '#/components/schemas/ItemsAssinaturaDTO'
                required: true
            responses:
                '200':
                    description: OK
                    content:
                        '*/*':
                            schema:
                                $ref: '#/components/schemas/ItemsAssinaturaDTO'
...
    /api/assinaturas/{id}:
        get:
            tags:
                - assinatura-resource
            operationId: getAssinatura
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                '200':
                    description: OK
                    content:
                        '*/*':
                            schema:
                                $ref: '#/components/schemas/AssinaturaDTO'
        put:
            tags:
                - assinatura-resource
            operationId: updateAssinatura
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AssinaturaDTO'
                required: true
            responses:
                '200':
                    description: OK
                    content:
                        '*/*':
                            schema:
                                $ref: '#/components/schemas/AssinaturaDTO'
        delete:
            tags:
                - assinatura-resource
            operationId: deleteAssinatura
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                '200':
                    description: OK
        patch:
            tags:
                - assinatura-resource
            operationId: partialUpdateAssinatura
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AssinaturaDTO'
                    application/merge-patch+json:
                        schema:
                            $ref: '#/components/schemas/AssinaturaDTO'
                required: true
            responses:
                '200':
                    description: OK
                    content:
                        '*/*':
                            schema:
                                $ref: '#/components/schemas/AssinaturaDTO'
...
    /api/items-assinaturas:
        get:
            tags:
                - items-assinatura-resource
            operationId: getAllItemsAssinaturas
            parameters:
                - name: page
                  in: query
                  description: Zero-based page index (0..N)
                  required: false
                  schema:
                      minimum: 0
                      type: integer
                      default: 0
                - name: size
                  in: query
                  description: The size of the page to be returned
                  required: false
                  schema:
                      minimum: 1
                      type: integer
                      default: 20
                - name: sort
                  in: query
                  description: >-
                      Sorting criteria in the format: property,(asc|desc). Default sort
                      order is ascending. Multiple sort criteria are supported.
                  required: false
                  schema:
                      type: array
                      items:
                          type: string
            responses:
                '200':
                    description: OK
                    content:
                        '*/*':
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/ItemsAssinaturaDTO'
        post:
            tags:
                - items-assinatura-resource
            operationId: createItemsAssinatura
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ItemsAssinaturaDTO'
                required: true
            responses:
                '200':
                    description: OK
                    content:
                        '*/*':
                            schema:
                                $ref: '#/components/schemas/ItemsAssinaturaDTO'
...
    /api/assinaturas:
        get:
            tags:
                - assinatura-resource
            operationId: getAllAssinaturas
            parameters:
                - name: page
                  in: query
                  description: Zero-based page index (0..N)
                  required: false
                  schema:
                      minimum: 0
                      type: integer
                      default: 0
                - name: size
                  in: query
                  description: The size of the page to be returned
                  required: false
                  schema:
                      minimum: 1
                      type: integer
                      default: 20
                - name: sort
                  in: query
                  description: >-
                      Sorting criteria in the format: property,(asc|desc). Default sort
                      order is ascending. Multiple sort criteria are supported.
                  required: false
                  schema:
                      type: array
                      items:
                          type: string
            responses:
                '200':
                    description: OK
                    content:
                        '*/*':
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/AssinaturaDTO'
        post:
            tags:
                - assinatura-resource
            operationId: createAssinatura
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AssinaturaDTO'
                required: true
            responses:
                '200':
                    description: OK
                    content:
                        '*/*':
                            schema:
                                $ref: '#/components/schemas/AssinaturaDTO'
...
components:
    schemas:
        PadariaDTO:
            required:
                - bairro
                - cep
                - cnpj
                - estado
                - fantasia
                - id
                - nome
                - numero
                - rua
                - telefone
            type: object
            properties:
                id:
                    minimum: 1
                    type: integer
                    format: int64
                nome:
                    maxLength: 255
                    minLength: 3
                    type: string
                fantasia:
                    maxLength: 255
                    minLength: 3
                    type: string
                telefone:
                    maxLength: 15
                    minLength: 10
                    type: string
                cnpj:
                    maxLength: 14
                    minLength: 14
                    type: string
                cep:
                    maxLength: 8
                    minLength: 8
                    type: string
                rua:
                    maxLength: 255
                    minLength: 3
                    type: string
                bairro:
                    maxLength: 255
                    minLength: 3
                    type: string
                estado:
                    maxLength: 2
                    minLength: 2
                    type: string
                numero:
                    maxLength: 255
                    minLength: 1
                    type: string
                complemento:
                    type: string
                contato:
                    type: string
                user:
                    $ref: '#/components/schemas/UserDTO'
        UserDTO:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                login:
                    type: string
        AssinaturaDTO:
            required:
                - ativa
                - id
                - nome
                - pagamentoRecorrenciaId
                - quantidadeDias
                - valor
            type: object
            properties:
                id:
                    minimum: 1
                    type: integer
                    format: int64
                valor:
                    type: number
                nome:
                    maxLength: 255
                    minLength: 3
                    type: string
                quantidadeDias:
                    minimum: 1
                    type: integer
                    format: int32
                ativa:
                    type: boolean
                pagamentoRecorrenciaId:
                    minimum: 1
                    type: integer
                    format: int64
                foto:
                    type: array
                    items:
                        type: string
                        format: byte
                fotoContentType:
                    type: string
                padaria:
                    $ref: '#/components/schemas/PadariaDTO'
                user:
                    $ref: '#/components/schemas/UserDTO'
        ItemsAssinaturaDTO:
            required:
                - id
                - nome
                - padariaId
                - valor
            type: object
            properties:
                id:
                    minimum: 1
                    type: integer
                    format: int64
                padariaId:
                    minimum: 1
                    type: integer
                    format: int64
                nome:
                    maxLength: 255
                    minLength: 3
                    type: string
                valor:
                    minimum: 1
                    type: number
                foto:
                    type: array
                    items:
                        type: string
                        format: byte
                fotoContentType:
                    type: string
                assinatura:
                    $ref: '#/components/schemas/AssinaturaDTO'
...